Add unit tests and README.
