+++ README.md
@@ -1,0 +1,16 @@
+# OpenVASP Java Core
+
+[![Open in Visual Studio Code](https://open.vscode.dev/badges/open-in-vscode.svg)](https://open.vscode.dev/OpenVASP/openvasp-java-client)
+
+This is a reference implementation of a Java core library for the OpenVASP standard.
+
+
+## What is OpenVASP?
+
+**VASP** stands for *Virtual Asset Service Provider*.
+OpenVASP is an open protocol to implement FATF's travel rule for virtual assets.
+
+The protocol facilitates robust compliance for VASPs, solely based on a set of principles, regardless of jurisdiction or virtual asset and without membership or registration with a centralized third-party.
+
+[Old Whitepaper](https://openvasp.org/wp-content/uploads/2020/09/OpenVasp_Whitepaper.pdf) | [Read more →](https://www.openvasp.org/)
+
