+++ src/main/java/org/openvasp/core/model/ivms101/Beneficiary.java
@@ -11,7 +11,7 @@
  */
 
 
-package org.openvasp.core.model;
+package org.openvasp.core.model.ivms101;
 
 import java.util.Objects;
 
@@ -22,57 +22,57 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Originator
+ * Beneficiary
  */
-@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-03-05T14:32:36.025401+02:00[Europe/Kiev]")
-public class Originator {
-  public static final String SERIALIZED_NAME_ORIGINATOR_PERSONS = "originatorPersons";
-  @SerializedName(SERIALIZED_NAME_ORIGINATOR_PERSONS)
-  private List<Person> originatorPersons = null;
+@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen")
+public class Beneficiary {
+  public static final String SERIALIZED_NAME_BENEFICIARY_PERSONS = "beneficiaryPersons";
+  @SerializedName(SERIALIZED_NAME_BENEFICIARY_PERSONS)
+  private List<Person> beneficiaryPersons = null;
 
   public static final String SERIALIZED_NAME_ACCOUNT_NUMBERS = "accountNumber";
   @SerializedName(SERIALIZED_NAME_ACCOUNT_NUMBERS)
   private List<String> accountNumbers = null;
 
 
-  public Originator originatorPersons(List<Person> originatorPersons) {
+  public Beneficiary beneficiaryPersons(List<Person> beneficiaryPersons) {
     
-    this.originatorPersons = originatorPersons;
+    this.beneficiaryPersons = beneficiaryPersons;
     return this;
   }
 
-  public Originator addOriginatorPersonsItem(Person originatorPersonsItem) {
-    if (this.originatorPersons == null) {
-      this.originatorPersons = new ArrayList<Person>();
+  public Beneficiary addBeneficiaryPersonsItem(Person beneficiaryPersonsItem) {
+    if (this.beneficiaryPersons == null) {
+      this.beneficiaryPersons = new ArrayList<Person>();
     }
-    this.originatorPersons.add(originatorPersonsItem);
+    this.beneficiaryPersons.add(beneficiaryPersonsItem);
     return this;
   }
 
    /**
-   * Get originatorPersons
-   * @return originatorPersons
+   * Get beneficiaryPersons
+   * @return beneficiaryPersons
   **/
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public List<Person> getOriginatorPersons() {
-    return originatorPersons;
+  public List<Person> getBeneficiaryPersons() {
+    return beneficiaryPersons;
   }
 
 
-  public void setOriginatorPersons(List<Person> originatorPersons) {
-    this.originatorPersons = originatorPersons;
+  public void setBeneficiaryPersons(List<Person> beneficiaryPersons) {
+    this.beneficiaryPersons = beneficiaryPersons;
   }
 
 
-  public Originator accountNumbers(List<String> accountNumbers) {
+  public Beneficiary accountNumbers(List<String> accountNumbers) {
     
     this.accountNumbers = accountNumbers;
     return this;
   }
 
-  public Originator addAccountNumbersItem(String accountNumbersItem) {
+  public Beneficiary addAccountNumbersItem(String accountNumbersItem) {
     if (this.accountNumbers == null) {
       this.accountNumbers = new ArrayList<String>();
     }
@@ -105,21 +105,21 @@ public class Originator {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    Originator originator = (Originator) o;
-    return Objects.equals(this.originatorPersons, originator.originatorPersons) &&
-        Objects.equals(this.accountNumbers, originator.accountNumbers);
+    Beneficiary beneficiary = (Beneficiary) o;
+    return Objects.equals(this.beneficiaryPersons, beneficiary.beneficiaryPersons) &&
+        Objects.equals(this.accountNumbers, beneficiary.accountNumbers);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(originatorPersons, accountNumbers);
+    return Objects.hash(beneficiaryPersons, accountNumbers);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class Originator {\n");
-    sb.append("    originatorPersons: ").append(toIndentedString(originatorPersons)).append("\n");
+    sb.append("class Beneficiary {\n");
+    sb.append("    beneficiaryPersons: ").append(toIndentedString(beneficiaryPersons)).append("\n");
     sb.append("    accountNumbers: ").append(toIndentedString(accountNumbers)).append("\n");
     sb.append("}");
     return sb.toString();
