+++ src/main/java/org/openvasp/core/model/ivms101/IntermediaryVasp.java
@@ -11,7 +11,7 @@
  */
 
 
-package org.openvasp.core.model;
+package org.openvasp.core.model.ivms101;
 
 import java.util.Objects;
 
@@ -21,7 +21,7 @@ import io.swagger.annotations.ApiModelProperty;
 /**
  * IntermediaryVasp
  */
-@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-03-05T14:32:36.025401+02:00[Europe/Kiev]")
+@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen")
 public class IntermediaryVasp {
   public static final String SERIALIZED_NAME_INTERMEDIARY_VASP = "intermediaryVASP";
   @SerializedName(SERIALIZED_NAME_INTERMEDIARY_VASP)
