+++ src/main/java/org/openvasp/core/model/ivms101/LegalPersonName.java
@@ -11,7 +11,7 @@
  */
 
 
-package org.openvasp.core.model;
+package org.openvasp.core.model.ivms101;
 
 import java.util.Objects;
 
@@ -24,7 +24,7 @@ import java.util.List;
 /**
  * LegalPersonName
  */
-@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-03-05T14:32:36.025401+02:00[Europe/Kiev]")
+@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen")
 public class LegalPersonName {
   public static final String SERIALIZED_NAME_NAME_IDENTIFIERS = "nameIdentifier";
   @SerializedName(SERIALIZED_NAME_NAME_IDENTIFIERS)
