+++ src/main/java/org/openvasp/core/model/ivms101/LegalPersonNameId.java
@@ -11,7 +11,7 @@
  */
 
 
-package org.openvasp.core.model;
+package org.openvasp.core.model.ivms101;
 
 import java.util.Objects;
 
@@ -26,7 +26,7 @@ import java.io.IOException;
 /**
  * LegalPersonNameId
  */
-@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-03-05T14:32:36.025401+02:00[Europe/Kiev]")
+@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen")
 public class LegalPersonNameId {
   public static final String SERIALIZED_NAME_LEGAL_PERSON_NAME = "legalPersonName";
   @SerializedName(SERIALIZED_NAME_LEGAL_PERSON_NAME)
